Move test base and change comment
